Lundberg vagrant

+++ start.sh
@@ -5,6 +5,16 @@ if [ ! -f eduid/compose.yml ]; then
     exit 1
 fi
 
+while [ "$1" != "" ]; do
+    case $1 in
+        --vagrant )             vagrant=true
+                                ;;
+        -h | --help )           echo "usage: start.sh [--vagrant]"
+                                exit
+    esac
+    shift
+done
+
 #
 # Set up entrys in /etc/hosts for the containers with externally accessible services
 # DON'T USE THE eduid_dev ONES, COOKIES ARE SCOPED FOR eduid.docker
@@ -45,6 +55,10 @@ fi
     fi
 done
 
-./bin/docker-compose -f eduid/compose.yml rm -s -f
-./bin/docker-compose -f eduid/compose.yml up $*
-./bin/docker-compose -f eduid/compose.yml logs -tf
+if [ "$vagrant" = true ]; then
+    vagrant ssh -c "cd /opt/eduid-developer; make start"
+else
+    ./bin/docker-compose -f eduid/compose.yml rm -s -f
+    ./bin/docker-compose -f eduid/compose.yml up $*
+    ./bin/docker-compose -f eduid/compose.yml logs -tf
+fi
